chore: removed unnecessary useMemo and moved dropDownItem interface to interface.ts
